Fix crash in CreateArmorNode (callee sub_1401CAFB0)
Error message:
Unhandled exception at 0x00007FF7F214B5B6 in SkyrimSE.exe: 0xC0000005: Access violation reading location 0x0000000000000130.

The faulting instruction is:
00000001401CB5B6  rdi,qword ptr [r12+130h]
which crashes due to a wrong NULL check for r12 above:
00000001401CB53A  test r12,r12
00000001401CB53D  jz short loc_1401CB5B6

This fix patches the jump so that the faulting block
is skipped if r12 is NULL.
This seems to be the originally intended behavior.

I have no idea what this function does.
The name CreateArmorNode comes from RaceMenu:
https://github.com/expired6978/SKSE64Plugins/blob/f6adf8cd14ffb65dba212ef85cc36093ac86051e/skee/SKEEHooks.cpp#L85
(Which originally appeared in the call stack when I was debugging this)